types: move mempool error for consistency
